GreenHub/momoapi/greenhub_trx_django/requirements.txt
django
djangorestframework

mtnmomoapi

psycopg2-binary

requests
ipykernel
python-dotenv